backend/pyproject.toml
[project]
name = "serenity-backend"
version = "1.0.0"
description = "Serenity salon booking platform backend"
requires-python = ">=3.12"
dependencies = [
    # Django core
    "django>=5.0.10,<5.1",
    "djangorestframework>=3.15.2,<4",
    "django-cors-headers>=4.6,<5",
    "django-filter>=24,<26",
    "django-csp>=4.0,<5",
    "django-ratelimit>=4.1,<5",

    # Database
    "psycopg2-binary>=2.9.10,<3",
    "dj-database-url>=2.3,<3",

    # Config
    "python-decouple>=3.8,<4",

    # Date / Time
    "python-dateutil>=2.9,<3",

    # Wagtail CMS
    "wagtail>=7.3,<8",
    "wagtail-localize>=1.10,<2",

    # Images
    "Pillow>=11.0,<12",

    # Cloudinary (media storage in production)
    "cloudinary>=1.40,<2",
    "django-cloudinary-storage>=0.3.0,<1",

    # Google Calendar integration
    "google-auth>=2.37,<3",
    "google-auth-oauthlib>=1.2,<2",
    "google-auth-httplib2>=0.2,<1",
    "google-api-python-client>=2.158,<3",

    # API docs
    "drf-spectacular>=0.28,<1",

    # Production server & static files
    "gunicorn>=23,<24",
    "whitenoise>=6.8,<7",
]

[project.optional-dependencies]
prod = [
    "django-redis>=5.4,<6",
    "hiredis>=3.0,<4",
]

# ── Dev dependency group (PEP 735) ──────────────────
[dependency-groups]
dev = [
    # Linting & formatting
    "ruff>=0.8.6",

    # Type checking
    "mypy>=1.13,<2",
    "django-stubs>=5.1,<6",
    "djangorestframework-stubs>=3.15,<4",
    "types-python-dateutil>=2.9,<3",
    "types-requests>=2.32,<3",

    # Dev tools
    "django-extensions>=3.2,<4",
    "ipython>=8.31,<9",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["apps*", "config*"]

# ── Mypy ────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
mypy_path = "."
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

# ── Strictness (start pragmatic, tighten over time) ─
strict = false
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true

# ── Import discovery ────────────────────────────────
ignore_missing_imports = false
namespace_packages = true
explicit_package_bases = true

# ── Per-module overrides ────────────────────────────

# Third-party libs without stubs / py.typed
[[tool.mypy.overrides]]
module = [
    "decouple",
    "decouple.*",
    "cloudinary.*",
    "cloudinary_storage.*",
    "wagtail.*",
    "wagtail_localize.*",
    "modelcluster.*",
    "taggit.*",
    "dj_database_url.*",
    "django_filters.*",
    "drf_spectacular.*",
    "csp.*",
    "django_ratelimit.*",
    "google.*",
    "googleapiclient.*",
    "httplib2.*",
    "pytest",
    "_pytest.*",
]
ignore_missing_imports = true

# Pragmatic ignores for noisy areas (tighten later)
[[tool.mypy.overrides]]
module = [
    "apps.*.migrations.*",
    "apps.*.tests.*",
    "apps.cms.*",
    "apps.testimonials.*",
    "config.settings",
    "config.settings.*",
]
ignore_errors = true

# ── django-stubs config ────────────────────────────
[tool.django-stubs]
django_settings_module = "config.settings"

# ── Ruff ────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*/migrations/*",
    "staticfiles",
    "media",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "DJ",   # flake8-django
    "ANN",  # flake8-annotations
    "TCH",  # flake8-type-checking
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (formatter handles it)
    "B008",   # function calls in argument defaults
    "B904",   # raise without from
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "ANN401", # Dynamically typed expressions (typing.Any) disallowed
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"config/settings/*.py" = ["F405", "F403", "ANN"]
"*/migrations/*.py" = ["ANN"]
"manage.py" = ["ANN"]
"**/tests/**.py" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "ANN204"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
